API read&write routes all implemented and well tested

+++ app_server/api/v1/views/folders.py
@@ -9,7 +9,7 @@ from db.db_classes import User, Folder
 
 @app_bp.route('/<user_id>/folders',
              methods=['GET', 'POST'], strict_slashes=False)
-def user_folders_viewer(user_id):
+def user_folders_route(user_id):
     user = main_storage.get(User, user_id)
     if user is None:
         return jsonify({'error': 'user not found'})
@@ -22,13 +22,14 @@ def user_folders_viewer(user_id):
             try:
                 folder = Folder(**folder_dict)
                 main_storage.new(folder)
+                main_storage.save()
                 return jsonify(folder.dict_repr())
             except OperationalError:
                 return jsonify({'error': 'json missing folder parameters'})
 
 @app_bp.route('/<user_id>/folders/<folder_id>',
-             methods=['PUT', 'DELETE'], strict_slashes=False)
-def user_folders_modifier(user_id, folder_id):
+              methods=['GET', 'PUT', 'DELETE'], strict_slashes=False)
+def user_folder_route(user_id, folder_id):
     user = main_storage.get(User, user_id)
     if user is None:
         return jsonify({'error': 'user not found'})
@@ -37,12 +38,17 @@ def user_folders_modifier(user_id, folder_id):
     if folder is None or folder.user_id != user.id:
         return jsonify({'error': 'folder not found'})
     match request.method:
+        case 'GET':
+            return jsonify(folder.dict_repr())
         case 'PUT':
             dic = request.get_json(silent=True)
             if type(dic) is dict:
                 folder.update(dic)
                 return jsonify(folder.dict_repr())
+            return jsonify({})
         case 'DELETE':
+            if folder_id == '00000000-0000-0000-0000-000000000000':
+                return jsonify({'error': f'cannot delete id {folder_id}'})
             main_storage.delete(folder)
             main_storage.save()
             return jsonify({})
